Refactor Accordion to use app-components pattern
- Created AccordionItem component in app-components for reusability
- Updated layout Accordion to use new app-component architecture
- Refactored LayoutSetSummaryAccordion to use app-components only
- Fixed nested Details structure bug in LayoutSetSummaryAccordion
- All existing tests continue to pass
