+++ src/layout/Summary2/CommonSummaryComponents/LayoutSetSummaryAccordion.tsx
@@ -1,7 +1,6 @@
 import React from 'react';
 
-import { Details, Label } from '@digdir/designsystemet-react';
-
+import { AccordionItem } from 'src/app-components/Accordion/AccordionItem';
 import { Flex } from 'src/app-components/Flex/Flex';
 import { Lang } from 'src/features/language/Lang';
 import classes from 'src/layout/Summary2/CommonSummaryComponents/LayoutSetSummaryAccordion.module.css';
@@ -14,35 +13,24 @@ type LayoutSetAccordionSummaryProps = {
 
 export function LayoutSetSummaryAccordion({ filteredPages }: LayoutSetAccordionSummaryProps) {
   return filteredPages.map((layoutId: string) => (
-    <Details
+    <AccordionItem
       key={layoutId}
       defaultOpen={true}
-      color='neutral'
       className={classes.summaryItem}
+      title={<Lang id={layoutId} />}
     >
-      <Details>
-        <Details.Summary>
-          <Label asChild>
-            <span>
-              <Lang id={layoutId} />
-            </span>
-          </Label>
-        </Details.Summary>
-        <Details.Content>
-          <Flex
-            container
-            spacing={6}
-            alignItems='flex-start'
-          >
-            <EmptyChildrenBoundary>
-              <PageSummary
-                pageId={layoutId}
-                key={layoutId}
-              />
-            </EmptyChildrenBoundary>
-          </Flex>
-        </Details.Content>
-      </Details>
-    </Details>
+      <Flex
+        container
+        spacing={6}
+        alignItems='flex-start'
+      >
+        <EmptyChildrenBoundary>
+          <PageSummary
+            pageId={layoutId}
+            key={layoutId}
+          />
+        </EmptyChildrenBoundary>
+      </Flex>
+    </AccordionItem>
   ));
 }
